+++ ergo_runtime/src/traits/value_by_content.rs
@@ -10,7 +10,7 @@ use grease::{
 /// A grease trait which identifies a value by its content.
 #[grease_trait]
 pub trait ValueByContent {
-    async fn value_by_content(self) -> Value;
+    async fn value_by_content(self, deep: bool) -> Value;
 }
 
 impl ValueByContent {
@@ -18,7 +18,7 @@ impl ValueByContent {
     pub fn add_impl<T: GreaseType + std::hash::Hash + Send + Sync + 'static>(traits: &mut Traits) {
         traits.add_impl_for_type::<T, ValueByContent>(grease_trait_impl! {
             impl<T: GreaseType + Send + Sync + std::hash::Hash + 'static> ValueByContent for T {
-                async fn value_by_content(self) -> Value {
+                async fn value_by_content(self, _deep: bool) -> Value {
                     let data = self.clone().await?;
                     let deps = depends![data.as_ref()];
                     Value::from(self).set_dependencies(deps)
@@ -28,7 +28,11 @@ impl ValueByContent {
     }
 }
 
-pub async fn value_by_content(ctx: &grease::runtime::Context, v: Value) -> grease::Result<Value> {
+pub async fn value_by_content(
+    ctx: &grease::runtime::Context,
+    v: Value,
+    deep: bool,
+) -> grease::Result<Value> {
     let t_ctx = ctx.clone();
     let mut t = ctx
         .get_trait::<ValueByContent, _, _>(&v, move |t| {
@@ -41,7 +45,7 @@ pub async fn value_by_content(ctx: &grease::runtime::Context, v: Value) -> greas
         })
         .await?;
 
-    t.value_by_content(v).await
+    t.value_by_content(v, deep).await
 }
 
 grease_traits_fn! {
@@ -51,16 +55,20 @@ grease_traits_fn! {
     ValueByContent::add_impl::<PathBuf>(traits);
 
     impl ValueByContent for types::Array {
-        async fn value_by_content(self) -> Value {
+        async fn value_by_content(self, deep: bool) -> Value {
             let data = self.await?;
             let types::Array(vals) = data.as_ref();
             let mut inner_vals = Vec::new();
             let mut errs: Vec<Error> = Vec::new();
-            for v in vals.iter() {
-                match super::value_by_content(CONTEXT, v.clone()).await {
-                    Ok(v) => inner_vals.push(v),
-                    Err(e) => errs.push(e)
+            if deep {
+                for v in vals.iter() {
+                    match super::value_by_content(CONTEXT, v.clone(), true).await {
+                        Ok(v) => inner_vals.push(v),
+                        Err(e) => errs.push(e)
+                    }
                 }
+            } else {
+                inner_vals = vals.iter().cloned().collect();
             }
             if !errs.is_empty() {
                 Err(errs.into_iter().collect::<Error>())?
@@ -70,16 +78,20 @@ grease_traits_fn! {
     }
 
     impl ValueByContent for types::Map {
-        async fn value_by_content(self) -> Value {
+        async fn value_by_content(self, deep: bool) -> Value {
             let data = self.await?;
             let types::Map(vals) = data.as_ref();
             let mut inner_vals = Vec::new();
             let mut errs: Vec<Error> = Vec::new();
-            for (k,v) in vals.iter() {
-                match super::value_by_content(CONTEXT, v.clone()).await {
-                    Ok(v) => inner_vals.push((k.clone(), v)),
-                    Err(e) => errs.push(e)
+            if deep {
+                for (k,v) in vals.iter() {
+                    match super::value_by_content(CONTEXT, v.clone(), true).await {
+                        Ok(v) => inner_vals.push((k.clone(), v)),
+                        Err(e) => errs.push(e)
+                    }
                 }
+            } else {
+                inner_vals = vals.iter().map(|(k,v)| (k.clone(), v.clone())).collect();
             }
             if !errs.is_empty() {
                 Err(errs.into_iter().collect::<Error>())?